requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Assets==2.1.0
webassets==2.0

# Database
PyMySQL==1.1.0
SQLAlchemy>=2.0.30
cryptography>=41.0.0

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0

# Payment
requests==2.31.0

# Email
Jinja2==3.1.2
MarkupSafe==2.1.3

# Asset processing
cssmin==0.2.0
jsmin==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Production
gunicorn==21.2.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
